+++ drivers/smbus/smbus_utils.h
