+++ tests/drivers/smbus/smbus_api/CMakeLists.txt
@@ -6,6 +6,7 @@ if(BOARD STREQUAL qemu_x86_64)
 	LIST(APPEND QEMU_EXTRA_FLAGS -machine q35)
 endif()
 
+target_include_directories(app PRIVATE ${ZEPHYR_BASE}/drivers/smbus)
 find_package(Zephyr REQUIRED HINTS $ENV{ZEPHYR_BASE})
 project(smbus_api)
